Refactor: project refactoring
#288
