+++ src/test/java/com/zerobase/foodlier/module/comment/reply/servcie/ReplyServiceTest.java
@@ -37,7 +37,7 @@ class ReplyServiceTest {
 
     @BeforeEach
     void setup() {
-        replyService = new ReplyServiceImpl(replyRepository);
+        replyService = new ReplyService(replyRepository);
     }
 
     @Test
